--- a/presto-parquet/src/main/resources/freemarker/config.fmpp
+++ b/presto-parquet/src/main/resources/freemarker/config.fmpp
@@ -1,0 +1,5 @@
+data: {
+    parquetTypes:   tdd(../data/ParquetTypes.tdd)
+    simpleTypeTemplate: text(../../../java/com/facebook/presto/parquet/batchreader/Int32FlatBatchReader.java)
+}
+