@zhenxiao 2020-05-22
is this related:
[ERROR] Tests run: 148, Failures: 1, Errors: 0, Skipped: 147, Time elapsed: 37.017 s <<< FAILURE! - in TestSuite
[ERROR] init(com.facebook.presto.hive.TestHivePushdownIntegrationSmokeTest)  Time elapsed: 36.416 s  <<< FAILURE!
java.lang.RuntimeException: 2
	at com.facebook.presto.tests.AbstractTestingPrestoClient.execute(AbstractTestingPrestoClient.java:115)
	at com.facebook.presto.tests.DistributedQueryRunner.execute(DistributedQueryRunner.java:436)
	at com.facebook.presto.tests.QueryAssertions.copyTable(QueryAssertions.java:364)
	at com.facebook.presto.tests.QueryAssertions.copyTable(QueryAssertions.java:356)
	at com.facebook.presto.tests.QueryAssertions.copyTpchTables(QueryAssertions.java:348)
	at com.facebook.presto.tests.QueryAssertions.copyTpchTables(QueryAssertions.java:334)
	at com.facebook.presto.hive.HiveQueryRunner.createQueryRunner(HiveQueryRunner.java:180)
	at com.facebook.presto.hive.HiveQueryRunner.createQueryRunner(HiveQueryRunner.java:110)
	at com.facebook.presto.hive.TestHivePushdownIntegrationSmokeTest.lambda$new$0(TestHivePushdownIntegrationSmokeTest.java:37)
	at com.facebook.presto.tests.AbstractTestQueryFramework.init(AbstractTestQueryFramework.java:85)
	at sun.reflect.NativeMethodAccessorImpl.invoke0(Native Method)
	at sun.reflect.NativeMethodAccessorImpl.invoke(NativeMethodAccessorImpl.java:62)
	at sun.reflect.DelegatingMethodAccessorImpl.invoke(DelegatingMethodAccessorImpl.java:43)
	at java.lang.reflect.Method.invoke(Method.java:498)
	at org.testng.internal.MethodInvocationHelper.invokeMethod(MethodInvocationHelper.java:104)
	at org.testng.internal.Invoker.invokeConfigurationMethod(Invoker.java:515)
	at org.testng.internal.Invoker.invokeConfigurations(Invoker.java:217)
	at org.testng.internal.Invoker.invokeConfigurations(Invoker.java:144)
	at org.testng.internal.TestMethodWorker.invokeBeforeClassMethods(TestMethodWorker.java:169)
	at org.testng.internal.TestMethodWorker.run(TestMethodWorker.java:108)
	at java.util.concurrent.ThreadPoolExecutor.runWorker(ThreadPoolExecutor.java:1149)
	at java.util.concurrent.ThreadPoolExecutor$Worker.run(ThreadPoolExecutor.java:624)
	at java.lang.Thread.run(Thread.java:748)
Caused by: java.lang.ArrayIndexOutOfBoundsException: 2
	at com.google.common.collect.RegularImmutableList.get(RegularImmutableList.java:60)
	at com.facebook.presto.execution.scheduler.FixedBucketNodeMap.getAssignedNode(FixedBucketNodeMap.java:44)
	at com.facebook.presto.execution.scheduler.BucketNodeMap.getAssignedNode(BucketNodeMap.java:48)
	at com.facebook.presto.execution.scheduler.NodeScheduler.selectDistributionNodes(NodeScheduler.java:303)
	at com.facebook.presto.execution.scheduler.nodeSelection.SimpleNodeSelector.computeAssignments(SimpleNodeSelector.java:199)
	at com.facebook.presto.execution.scheduler.FixedSourcePartitionedScheduler$BucketedSplitPlacementPolicy.computeAssignments(FixedSourcePartitionedScheduler.java:316)
	at com.facebook.presto.execution.scheduler.SourcePartitionedScheduler.schedule(SourcePartitionedScheduler.java:273)
	at com.facebook.presto.execution.scheduler.FixedSourcePartitionedScheduler.schedule(FixedSourcePartitionedScheduler.java:240)
	at com.facebook.presto.execution.scheduler.LegacySqlQueryScheduler.schedule(LegacySqlQueryScheduler.java:348)
	at java.util.concurrent.Executors$RunnableAdapter.call(Executors.java:511)
	at java.util.concurrent.FutureTask.run(FutureTask.java:266)
	... 3 more

[INFO] 
[INFO] Results:
[INFO] 
[ERROR] Failures: 
[ERROR]   TestHivePushdownIntegrationSmokeTest>AbstractTestQueryFramework.init:85->lambda$new$0:37 Â» Runtime
[INFO] 
[ERROR] Tests run: 148, Failures: 1, Errors: 0, Skipped: 147
[INFO]