@zhenxiao 2020-05-22
hi @vkorukanti currently we will merge individual commits in a PR as separate commits into master. Would be nice if we could keep each commit clean. Could you please resolve file merge in: https://github.com/prestodb/presto/pull/13638/commits/ddc8a6688a53cc1669fe48feb75c1775ce332e9f