@vkorukanti 2020-05-23
> hi @vkorukanti currently we will merge individual commits in a PR as separate commits into master. Would be nice if we could keep each commit clean. Could you please resolve file merge in: [ddc8a66](https://github.com/prestodb/presto/commit/ddc8a6688a53cc1669fe48feb75c1775ce332e9f)
Just force pushed, this should be fixed. The failed test ran fine in locally `mvn test -Ptest-hive-pushdown-filter-queries-basic`. Also it passed in latest CI run.